pyproject.toml
[tool.poetry]
name = "frmatcher"
version = "0.1.0"
description = ""
authors = ["Denis Odinokov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.7.2"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ipykernel = "^6.29.5"
ruff = "^0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
